requirements.txt
requests==2.25.1
prometheus_flask_exporter==0.18.7
psycopg2-binary==2.9.3